ETCM-1058 Add new Consensus, by just moving the code in BlockImport
